refactor(cli): extract the product linter's markdown primitives
`scripts/lib/product-lint.js` sat at 699 of the 700-line hard limit in
`.claude/rules/coding-standards.md`, so the next C-rule could not land in
it. `docs/plans/check-product-deferred.md` §4 prescribed this split; the
duplicate-`Version` clause in the next commit is the rule that forces it.

`FENCE_RE`, `section()`, `unfenced()` and `stripCodeSpans()` move verbatim
into `scripts/lib/product-markdown.js`. They know nothing about product
state — no roadmap row, no `D-id`, no spec header reaches them — and
nothing outside `product-lint.js` called them. Every rule, parser,
`validateProduct` and the exports stay put, so `scripts/check-product.js`
and `tests/scripts/check-product.test.js` import exactly what they did
before; `product-lint.js` comes out at 617 lines.

No behaviour change: the 108-case falsifiability suite passes unmoved,
and `check:product` is green on the live corpus.
